requirements.txt
streamlit
ollama
python-docx
pandas
python-dotenv
plotly